+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog: Reexpress Model-Context-Protocol (MCP) Server
 
+## What's new in version 2.5.0
+
+The version 2.5.0 release implements nested estimators as described in [Research Note: Nested Similarity-Distance-Magnitude Estimators](https://raw.githubusercontent.com/ReexpressAI/sdm_activations/main/research_notes/nested_sdm_estimators.pdf). This approach simply runs the calibration algorithm for descending probability thresholds on the class- and prediction-condition accuracy. The result is that rather than a single region, the vast majority of calibration points can be assigned to a region with a class- and prediction-condition accuracy estimated to be at least some value greater than 0.5. The remaining points not assigned to any region can be viewed as effectively out-of-distribution. The most conservative region retains its interpretation and behavior as before, and we find that the nested regions provide a meaningful ranking for adjudicating the relative probability of the remaining points. 
+
+Additionally, the code base has been streamlined, with the language model post-training code removed. A separate repo will be released for fine-training the underlying weights of networks.
+
+The released model is otherwise the same as version 2.4.x. It is calibrated using the same data and with gpt-5.5-2026-04-23 and gemini-3.1-pro-preview as the generative models. Refer to the [Version 2.4.0 model card](documentation/model_cards/model_card_v240.pdf) for details.
+
 ## What's new in version 2.4.1
 
 The version 2.4.1 release is identical to version 2.4.0 (including the model), but ensures the optional HTML output uses the SDM_HR^lower estimator for the confidence bin label at the top of the page when constants.MCP_SERVER_USE_DKW_LOWER_ESTIMATES is True. This does not impact the output to the tool-calling LM (or the estimator details elsewhere in the generated HTML).